Took the long way but fixed it
I´ve now successfully made the API convert markdown files to json and spit it out on the page. To do this I had to use two another converters; markdown (markdown to HTML) and html2json (HTML to JSON). This was because the first and direct md-2-json could only handle an h(x) having an h(x-1) inside it, not and h(x-2) or h(x-3) (an h2 having an h4 without an h3 inbetween would fail to be converted).