Now taking put-requests
The API can now successfully take and respond to put-requests coming from the frontend. This works on localhost in chrome but not in firefox. Don´t know why, but it has to do with CORS somehow. This commit is built for running locally, so the public build will not work.